Changed motor_voltage to 32 bit to test

+++ march_hardware/src/IMotionCube.cpp
@@ -268,7 +268,7 @@ float IMotionCube::getIMCVoltage()
 
 float IMotionCube::getMotorVoltage()
 {
-  uint16_t motor_voltage = 0;
+  uint32_t motor_voltage = 0;
   int val_size = 16;
   sdo_bit32_write(slaveIndex, 0x2064, 0, 0x0232);
   sdo_bit32_read(slaveIndex, 0x2066, 0, val_size, motor_voltage);
